+++ SonataPageBundle.php
@@ -14,6 +14,7 @@ use Symfony\Component\HttpKernel\Bundle\Bundle;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
 use Sonata\PageBundle\DependencyInjection\Compiler\CacheCompilerPass;
+use Sonata\PageBundle\DependencyInjection\Compiler\GlobalVariablesCompilerPass;
 
 class SonataPageBundle extends Bundle
 {
@@ -22,5 +23,6 @@ class SonataPageBundle extends Bundle
         parent::build($container);
 
         $container->addCompilerPass(new CacheCompilerPass());
+        $container->addCompilerPass(new GlobalVariablesCompilerPass());
     }
 }