+++ CHANGELOG.md
@@ -10,6 +10,9 @@ A [BC BREAK] means the update will break the project for many reasons :
 ### 2012-03-19
 
 * [BC BREAK] DecoratorStrategy : Update regular expressions, regular expression separators are not fixed in the class.
+* Add a twig global variable sonata_page (sonata_page.cmsmanager, sonata_page.siteavailables, sonata_page.currentsite)
+* Add locale to Site model, if the locale is set an _locale attribute is set to the request and no site available for
+  the current url, then the user is redirected to the default site a (ie, / => /en, if /en is the default)
 
 ### 2012-02-28